+++ changelog.rst
@@ -41,6 +41,7 @@ Last change: 09-APR-2026 CEST
 - Trace shows number of photons in addition to x, y and frame; exports .csv files with three columns (frame, ON/OFF and photons)
 - Render GUI: plot localization profile for rectangular pick
 - 3D rotation window supports rendering by property
+- Render, Average and Filter allow the user to inspect metadata in the app
 
 *Bug fixes:*
 ++++++++++++